fix: removed fullscreen to feature
