+++ packages/client/src/features/fullscreen/index.ts
@@ -1,0 +1,1 @@
+export { FullscreenButton } from './view'
