+++ packages/client/src/pages/login/style.css
@@ -1,7 +1,6 @@
 .login-page {
     width: 100%;
     height: 100%;
-
     display: flex;
     flex-direction: column;
     align-items: center;
